data_quality_pipeline/pyproject.toml
[project]
name = "made"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Leonardo Baroncelli", email = "leonardo.baroncelli.ext@leonardo.com" }
]
requires-python = ">=3.10"
dependencies = [
    "numpy<2",
    "ray[data,default]>=2.40.0",
    "webdataset",
    "transformers>=4.29.0",
    "pytest",
    "debugpy==1.8.0",
]

[project.optional-dependencies]
unimodal-text-filtering = [
    "nltk",
    "fasttext>=0.9.2"
]
unimodal-image-filtering = [
    "pillow>=11.1.0"
]
multimodal-filtering = [

]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
post-build = ["bash download_models.sh"]